refresh administrator roles after edit own profile
