+++ packages/framework/src/Model/Administrator/Security/AdministratorRolesChangedSubscriber.php
@@ -5,12 +5,10 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Model\Administrator\Security;
 
 use Shopsys\FrameworkBundle\Model\Administrator\Administrator;
-use Shopsys\FrameworkBundle\Model\Administrator\AdministratorFacade;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 
 class AdministratorRolesChangedSubscriber implements EventSubscriberInterface
 {
@@ -25,18 +23,20 @@ class AdministratorRolesChangedSubscriber implements EventSubscriberInterface
     protected $tokenStorage;
 
     /**
-     * @var \Shopsys\FrameworkBundle\Model\Administrator\AdministratorFacade
+     * @var \Shopsys\FrameworkBundle\Model\Administrator\Security\AdministratorRolesChangedFacade
      */
-    protected $administratorFacade;
+    protected $administratorRolesChangedFacade;
 
     /**
      * @param \Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface $tokenStorage
-     * @param \Shopsys\FrameworkBundle\Model\Administrator\AdministratorFacade $administratorFacade
+     * @param \Shopsys\FrameworkBundle\Model\Administrator\Security\AdministratorRolesChangedFacade $administratorRolesChangedFacade
      */
-    public function __construct(TokenStorageInterface $tokenStorage, AdministratorFacade $administratorFacade)
-    {
+    public function __construct(
+        TokenStorageInterface $tokenStorage,
+        AdministratorRolesChangedFacade $administratorRolesChangedFacade
+    ) {
         $this->tokenStorage = $tokenStorage;
-        $this->administratorFacade = $administratorFacade;
+        $this->administratorRolesChangedFacade = $administratorRolesChangedFacade;
     }
 
     /**
@@ -63,10 +63,7 @@ class AdministratorRolesChangedSubscriber implements EventSubscriberInterface
         }
 
         if ($administrator instanceof Administrator && $this->rolesChanged === true) {
-            $token = new UsernamePasswordToken($administrator, $administrator->getPassword(), 'administration', $administrator->getRoles());
-            $this->tokenStorage->setToken($token);
-            $this->administratorFacade->setRolesChangedNow($administrator);
-            $this->rolesChanged = false;
+            $this->administratorRolesChangedFacade->refreshAdministratorToken($administrator);
         }
     }
